add auth protection to routes

+++ backend/controllers/books.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from 'express';
 import asyncWrapper from '../middleware/async-wrapper';
 import { PrismaClient } from '@prisma/client';
+import { ROLE_FLAGS } from '../utils/roles';
+
 
 const prisma = new PrismaClient();
 
@@ -15,6 +17,25 @@ export const getAllBooks = asyncWrapper(async (req: Request, res: Response) => {
 })
 
 export const postCreateNewBook = asyncWrapper(async (req: Request, res: Response) => {
+    const authUserID = req.userID?.id
+    const authUserData = await prisma.user.findFirst({
+        where: {
+            id: authUserID
+        },
+        select: {
+            Role: true
+        }
+    })
+
+    // exclude 0 as 0 is a possible value for permissions
+    if (!authUserData?.Role?.permissions && authUserData?.Role?.permissions !== 0) {
+        return res.status(404).json({ error: "Ivalid token\\User not found" })
+    }
+
+    if ((authUserData?.Role?.permissions & ROLE_FLAGS.WRITE) !== ROLE_FLAGS.WRITE) {
+        return res.status(403).json({ error: "You do not have the permision" })
+    }
+
     const newBook = await prisma.book.create({
         data: {
             title: req.body.title,
@@ -36,6 +57,25 @@ export const getBookById = asyncWrapper(async (req: Request, res: Response) => {
 })
 
 export const putUpdateBookByID = asyncWrapper(async (req: Request, res: Response) => {
+    const authUserID = req.userID?.id
+    const authUserData = await prisma.user.findFirst({
+        where: {
+            id: authUserID
+        },
+        select: {
+            Role: true
+        }
+    })
+
+    // exclude 0 as 0 is a possible value for permissions
+    if (!authUserData?.Role?.permissions && authUserData?.Role?.permissions !== 0) {
+        return res.status(404).json({ error: "Ivalid token\\User not found" })
+    }
+
+    if ((authUserData?.Role?.permissions & ROLE_FLAGS.MODIFY) !== ROLE_FLAGS.MODIFY) {
+        return res.status(403).json({ error: "You do not have the permision" })
+    }
+
     const updatedBook = await prisma.book.update({
         where: {
             id: req.params.id
@@ -48,10 +88,29 @@ export const putUpdateBookByID = asyncWrapper(async (req: Request, res: Response
 })
 
 export const deleteBookByID = asyncWrapper(async (req: Request, res: Response) => {
+    const authUserID = req.userID?.id
+    const authUserData = await prisma.user.findFirst({
+        where: {
+            id: authUserID
+        },
+        select: {
+            Role: true
+        }
+    })
+
+    // exclude 0 as 0 is a possible value for permissions
+    if (!authUserData?.Role?.permissions && authUserData?.Role?.permissions !== 0) {
+        return res.status(404).json({ error: "Ivalid token\\User not found" })
+    }
+
+    if ((authUserData?.Role?.permissions & ROLE_FLAGS.DELETE) !== ROLE_FLAGS.DELETE) {
+        return res.status(403).json({ error: "You do not have the permision" })
+    }
+
     await prisma.book.delete({
         where: {
             id: req.params.id
         }
     })
-    res.status(201).send({success: true})
+    res.status(201).send({ success: true })
 })
